# fern/apis/prod/openapi/openapi.yaml
openapi: 3.0.2
info:
  title: Superagent
  description: The Open Source AI Assistant Framework & API
  version: 0.2.6
servers:
  - url: https://api.pixx.co
paths:
...
  /api/v1/api-users:
    post:
      tags:
        - Api user
      summary: Create
      description: Create a new API user
      operationId: create_api_v1_api_users_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/app__models__request__ApiUser'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/app__models__response__ApiUser'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/api-users/me:
    get:
      tags:
        - Api user
      summary: Get
      description: Get a single api user
      operationId: get_api_v1_api_users_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/app__models__response__ApiUser'
      security:
        - HTTPBearer: []
    delete:
      tags:
        - Api user
      summary: Delete
      description: Delete an api user
      operationId: delete_api_v1_api_users_me_delete
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
        - HTTPBearer: []
  /api/v1/api-users/identify:
    post:
      tags:
        - Api user
      summary: Indentify
      description: Indentify an api user
      operationId: indentify_api_v1_api_users_identify_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/app__models__request__ApiUser'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - HTTPBearer: []
...
components:
  schemas:
...
    AgentType:
      title: AgentType
      enum:
        - SUPERAGENT
        - OPENAI_ASSISTANT
      type: string
      description: An enumeration.
...
    DatasourceStatus:
      title: DatasourceStatus
      enum:
        - IN_PROGRESS
        - DONE
        - FAILED
      type: string
      description: An enumeration.
    DatasourceType:
      title: DatasourceType
      enum:
        - TXT
        - PDF
        - CSV
        - PPTX
        - XLSX
        - DOCX
        - GOOGLE_DOC
        - YOUTUBE
        - GITHUB_REPOSITORY
        - MARKDOWN
        - WEBPAGE
        - AIRTABLE
        - STRIPE
        - NOTION
        - SITEMAP
        - URL
        - FUNCTION
      type: string
      description: An enumeration.
...
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
...
    LLMModel:
      title: LLMModel
      enum:
        - GPT_3_5_TURBO_16K_0613
        - GPT_3_5_TURBO_0613
        - GPT_3_5_TURBO_1106
        - GPT_3_5_TURBO_0125
        - GPT_4_0613
        - GPT_4_32K_0613
        - GPT_4_1106_PREVIEW
        - GPT_4_TURBO_PREVIEW
        - MISTRAL_7B_INSTRUCT_V01
        - MIXTRAL_8X7B_INSTRUCT_V01
      type: string
      description: An enumeration.
...
    LLMProvider:
      title: LLMProvider
      enum:
        - OPENAI
        - AZURE_OPENAI
        - HUGGINGFACE
        - OPENROUTER
      type: string
      description: An enumeration.
...
    ToolType:
      title: ToolType
      enum:
        - ALGOLIA
        - BROWSER
        - BING_SEARCH
        - REPLICATE
        - WOLFRAM_ALPHA
        - ZAPIER_NLA
        - AGENT
        - OPENAPI
        - CHATGPT_PLUGIN
        - METAPHOR
        - PUBMED
        - CODE_EXECUTOR
        - OPENBB
        - GPT_VISION
        - TTS_1
        - HAND_OFF
        - FUNCTION
        - HTTP
      type: string
      description: An enumeration.
...
    ValidationError:
      title: ValidationError
      required:
        - loc
        - msg
        - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
...
    VectorDbProvider:
      title: VectorDbProvider
      enum:
        - PINECONE
        - ASTRA_DB
        - WEAVIATE
        - QDRANT
        - SUPABASE
      type: string
      description: An enumeration.
    WorkflowConfig:
      title: WorkflowConfig
      required:
        - id
        - createdAt
        - updatedAt
        - workflowId
      type: object
      properties:
        id:
          title: Id
          type: string
        config:
          title: Config
          type: string
          format: json-string
        createdAt:
          title: Createdat
          type: string
          format: date-time
        updatedAt:
          title: Updatedat
          type: string
          format: date-time
        workflowId:
          title: Workflowid
          type: string
        workflow:
          $ref: '#/components/schemas/prisma__models__Workflow'
        ApiUser:
          $ref: '#/components/schemas/prisma__models__ApiUser'
        apiUserId:
          title: Apiuserid
          type: string
      description: Represents a WorkflowConfig record
...
    app__models__request__ApiUser:
      title: ApiUser
      required:
        - email
      type: object
      properties:
        email:
          title: Email
          type: string
        firstName:
          title: Firstname
          type: string
        lastName:
          title: Lastname
          type: string
        company:
          title: Company
          type: string
        anonymousId:
          title: Anonymousid
          type: string
...
    app__models__response__ApiUser:
      title: ApiUser
      required:
        - success
      type: object
      properties:
        success:
          title: Success
          type: boolean
        data:
          $ref: '#/components/schemas/prisma__models__ApiUser'
...
    prisma__models__Agent:
      title: Agent
      required:
        - id
        - type
        - name
        - description
        - isActive
        - createdAt
        - updatedAt
        - apiUserId
      type: object
      properties:
        id:
          title: Id
          type: string
        type:
          $ref: '#/components/schemas/AgentType'
        name:
          title: Name
          type: string
        avatar:
          title: Avatar
          type: string
        initialMessage:
          title: Initialmessage
          type: string
        description:
          title: Description
          type: string
        isActive:
          title: Isactive
          type: boolean
        createdAt:
          title: Createdat
          type: string
          format: date-time
        updatedAt:
          title: Updatedat
          type: string
          format: date-time
        llms:
          title: Llms
          type: array
          items:
            $ref: '#/components/schemas/prisma__models__AgentLLM'
        llmModel:
          $ref: '#/components/schemas/LLMModel'
        prompt:
          title: Prompt
          type: string
        apiUserId:
          title: Apiuserid
          type: string
        apiUser:
          $ref: '#/components/schemas/prisma__models__ApiUser'
        datasources:
          title: Datasources
          type: array
          items:
            $ref: '#/components/schemas/prisma__models__AgentDatasource'
        tools:
          title: Tools
          type: array
          items:
            $ref: '#/components/schemas/prisma__models__AgentTool'
        workflowSteps:
          title: Workflowsteps
          type: array
          items:
            $ref: '#/components/schemas/prisma__models__WorkflowStep'
        metadata:
          title: Metadata
          type: string
          format: json-string
      description: Represents a Agent record
    prisma__models__AgentDatasource:
      title: AgentDatasource
      required:
        - agentId
        - datasourceId
        - createdAt
        - updatedAt
      type: object
      properties:
        agentId:
          title: Agentid
          type: string
        datasourceId:
          title: Datasourceid
          type: string
        agent:
          $ref: '#/components/schemas/prisma__models__Agent'
        datasource:
          $ref: '#/components/schemas/prisma__models__Datasource'
        createdAt:
          title: Createdat
          type: string
          format: date-time
        updatedAt:
          title: Updatedat
          type: string
          format: date-time
      description: Represents a AgentDatasource record
    prisma__models__AgentLLM:
      title: AgentLLM
      required:
        - agentId
        - llmId
        - createdAt
        - updatedAt
      type: object
      properties:
        agentId:
          title: Agentid
          type: string
        llmId:
          title: Llmid
          type: string
        agent:
          $ref: '#/components/schemas/prisma__models__Agent'
        llm:
          $ref: '#/components/schemas/prisma__models__LLM'
        createdAt:
          title: Createdat
          type: string
          format: date-time
        updatedAt:
          title: Updatedat
          type: string
          format: date-time
      description: Represents a AgentLLM record
    prisma__models__AgentTool:
      title: AgentTool
      required:
        - agentId
        - toolId
        - createdAt
        - updatedAt
      type: object
      properties:
        agentId:
          title: Agentid
          type: string
        toolId:
          title: Toolid
          type: string
        agent:
          $ref: '#/components/schemas/prisma__models__Agent'
        tool:
          $ref: '#/components/schemas/prisma__models__Tool'
        createdAt:
          title: Createdat
          type: string
          format: date-time
        updatedAt:
          title: Updatedat
          type: string
          format: date-time
      description: Represents a AgentTool record
    prisma__models__ApiUser:
      title: ApiUser
      required:
        - id
        - createdAt
        - updatedAt
      type: object
      properties:
        id:
          title: Id
          type: string
        token:
          title: Token
          type: string
        email:
          title: Email
          type: string
        createdAt:
          title: Createdat
          type: string
          format: date-time
        updatedAt:
          title: Updatedat
          type: string
          format: date-time
        agents:
          title: Agents
          type: array
          items:
            $ref: '#/components/schemas/prisma__models__Agent'
        llms:
          title: Llms
          type: array
          items:
            $ref: '#/components/schemas/prisma__models__LLM'
        datasources:
          title: Datasources
          type: array
          items:
            $ref: '#/components/schemas/prisma__models__Datasource'
        tools:
          title: Tools
          type: array
          items:
            $ref: '#/components/schemas/prisma__models__Tool'
        workflows:
          title: Workflows
          type: array
          items:
            $ref: '#/components/schemas/prisma__models__Workflow'
        vectorDb:
          title: Vectordb
          type: array
          items:
            $ref: '#/components/schemas/prisma__models__VectorDb'
        workflowConfigs:
          title: Workflowconfigs
          type: array
          items:
            $ref: '#/components/schemas/WorkflowConfig'
      description: Represents a ApiUser record
    prisma__models__Datasource:
      title: Datasource
      required:
        - id
        - name
        - type
        - apiUserId
        - createdAt
        - updatedAt
        - status
      type: object
      properties:
        id:
          title: Id
          type: string
        name:
          title: Name
          type: string
        content:
          title: Content
          type: string
        description:
          title: Description
          type: string
        url:
          title: Url
          type: string
        type:
          $ref: '#/components/schemas/DatasourceType'
        apiUserId:
          title: Apiuserid
          type: string
        apiUser:
          $ref: '#/components/schemas/prisma__models__ApiUser'
        createdAt:
          title: Createdat
          type: string
          format: date-time
        updatedAt:
          title: Updatedat
          type: string
          format: date-time
        metadata:
          title: Metadata
          type: string
        status:
          $ref: '#/components/schemas/DatasourceStatus'
        datasources:
          title: Datasources
          type: array
          items:
            $ref: '#/components/schemas/prisma__models__AgentDatasource'
        vectorDb:
          $ref: '#/components/schemas/prisma__models__VectorDb'
        vectorDbId:
          title: Vectordbid
          type: string
      description: Represents a Datasource record
    prisma__models__LLM:
      title: LLM
      required:
        - id
        - provider
        - apiKey
        - createdAt
        - updatedAt
        - apiUserId
      type: object
      properties:
        id:
          title: Id
          type: string
        provider:
          $ref: '#/components/schemas/LLMProvider'
        apiKey:
          title: Apikey
          type: string
        options:
          title: Options
          type: string
          format: json-string
        agents:
          title: Agents
          type: array
          items:
            $ref: '#/components/schemas/prisma__models__AgentLLM'
        createdAt:
          title: Createdat
          type: string
          format: date-time
        updatedAt:
          title: Updatedat
          type: string
          format: date-time
        apiUserId:
          title: Apiuserid
          type: string
        apiUser:
          $ref: '#/components/schemas/prisma__models__ApiUser'
      description: Represents a LLM record
    prisma__models__Tool:
      title: Tool
      required:
        - id
        - name
        - description
        - type
        - returnDirect
        - createdAt
        - updatedAt
        - apiUserId
      type: object
      properties:
        id:
          title: Id
          type: string
        name:
          title: Name
          type: string
        description:
          title: Description
          type: string
        type:
          $ref: '#/components/schemas/ToolType'
        returnDirect:
          title: Returndirect
          type: boolean
        metadata:
          title: Metadata
          type: string
        createdAt:
          title: Createdat
          type: string
          format: date-time
        updatedAt:
          title: Updatedat
          type: string
          format: date-time
        apiUserId:
          title: Apiuserid
          type: string
        apiUser:
          $ref: '#/components/schemas/prisma__models__ApiUser'
        tools:
          title: Tools
          type: array
          items:
            $ref: '#/components/schemas/prisma__models__AgentTool'
        toolConfig:
          title: Toolconfig
          type: string
          format: json-string
      description: Represents a Tool record
    prisma__models__VectorDb:
      title: VectorDb
      required:
        - id
        - provider
        - createdAt
        - updatedAt
        - apiUserId
      type: object
      properties:
        id:
          title: Id
          type: string
        provider:
          $ref: '#/components/schemas/VectorDbProvider'
        options:
          title: Options
          type: string
          format: json-string
        datasources:
          title: Datasources
          type: array
          items:
            $ref: '#/components/schemas/prisma__models__Datasource'
        createdAt:
          title: Createdat
          type: string
          format: date-time
        updatedAt:
          title: Updatedat
          type: string
          format: date-time
        apiUserId:
          title: Apiuserid
          type: string
        apiUser:
          $ref: '#/components/schemas/prisma__models__ApiUser'
      description: Represents a VectorDb record
    prisma__models__Workflow:
      title: Workflow
      required:
        - id
        - name
        - createdAt
        - updatedAt
        - apiUserId
      type: object
      properties:
        id:
          title: Id
          type: string
        name:
          title: Name
          type: string
        description:
          title: Description
          type: string
        createdAt:
          title: Createdat
          type: string
          format: date-time
        updatedAt:
          title: Updatedat
          type: string
          format: date-time
        steps:
          title: Steps
          type: array
          items:
            $ref: '#/components/schemas/prisma__models__WorkflowStep'
        apiUserId:
          title: Apiuserid
          type: string
        apiUser:
          $ref: '#/components/schemas/prisma__models__ApiUser'
        workflowConfigs:
          title: Workflowconfigs
          type: array
          items:
            $ref: '#/components/schemas/WorkflowConfig'
      description: Represents a Workflow record
    prisma__models__WorkflowStep:
      title: WorkflowStep
      required:
        - id
        - order
        - workflowId
        - createdAt
        - updatedAt
        - agentId
      type: object
      properties:
        id:
          title: Id
          type: string
        order:
          title: Order
          type: integer
        workflowId:
          title: Workflowid
          type: string
        workflow:
          $ref: '#/components/schemas/prisma__models__Workflow'
        createdAt:
          title: Createdat
          type: string
          format: date-time
        updatedAt:
          title: Updatedat
          type: string
          format: date-time
        input:
          title: Input
          type: string
        output:
          title: Output
          type: string
        agentId:
          title: Agentid
          type: string
        agent:
          $ref: '#/components/schemas/prisma__models__Agent'
      description: Represents a WorkflowStep record
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer